@@ -1,14 +1,19 @@
 // Copyright 2022-2024, Offchain Labs, Inc.
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
-//go:build js
-// +build js
+//go:build wasm
+// +build wasm
 
 package programs
 
 import (
+	"encoding/binary"
+	"errors"
+	"unsafe"
+
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
+	"github.com/ethereum/go-ethereum/log"
 	"github.com/offchainlabs/nitro/arbos/burn"
 	"github.com/offchainlabs/nitro/arbos/util"
 	"github.com/offchainlabs/nitro/arbutil"
@@ -31,32 +36,20 @@ type rustConfig byte
 type rustModule byte
 type rustEvmData byte
 
-func activateProgramRustImpl(
-	wasm []byte, pageLimit, version u16, debugMode u32, moduleHash *hash, gas *u64,
-) (initGas, asmEstimate u32, footprint u16, err *rustVec)
-
-func callProgramRustImpl(
-	moduleHash *hash, calldata []byte, params *rustConfig, evmApi uint32, evmData *rustEvmData, gas *u64,
-) (status userStatus, out *rustVec)
-
-func readRustVecLenImpl(vec *rustVec) (len u32)
-func rustVecIntoSliceImpl(vec *rustVec, ptr *byte)
-func rustModuleDropImpl(mach *rustModule)
-func rustConfigImpl(version u16, maxDepth, inkPrice, debugMode u32) *rustConfig
-func rustEvmDataImpl(
-	blockBasefee *hash,
-	chainId u64,
-	blockCoinbase *addr,
-	blockGasLimit u64,
-	blockNumber u64,
-	blockTimestamp u64,
-	contractAddress *addr,
-	msgSender *addr,
-	msgValue *hash,
-	txGasPrice *hash,
-	txOrigin *addr,
-	reentrant u32,
-) *rustEvmData
+//go:wasmimport programs activate
+func programActivate(
+	wasm_ptr unsafe.Pointer,
+	wasm_size uint32,
+	pages_ptr unsafe.Pointer,
+	asm_estimation_ptr unsafe.Pointer,
+	init_gas_ptr unsafe.Pointer,
+	version uint32,
+	debug uint32,
+	module_hash_ptr unsafe.Pointer,
+	gas_ptr unsafe.Pointer,
+	err_buf unsafe.Pointer,
+	err_buf_len uint32,
+) uint32
 
 func activateProgram(
 	db vm.StateDB,
@@ -67,19 +60,61 @@ func activateProgram(
 	debug bool,
 	burner burn.Burner,
 ) (*activationInfo, error) {
+	errBuf := make([]byte, 1024)
 	debugMode := arbmath.BoolToUint32(debug)
 	moduleHash := common.Hash{}
 	gasPtr := burner.GasLeft()
+	asmEstimate := uint32(0)
+	initGas := uint32(0)
 
-	initGas, asmEstimate, footprint, err := activateProgramRustImpl(
-		wasm, pageLimit, version, debugMode, &moduleHash, gasPtr,
+	footprint := uint16(pageLimit)
+	errLen := programActivate(
+		arbutil.SliceToUnsafePointer(wasm),
+		uint32(len(wasm)),
+		unsafe.Pointer(&footprint),
+		unsafe.Pointer(&asmEstimate),
+		unsafe.Pointer(&initGas),
+		uint32(version),
+		debugMode,
+		arbutil.SliceToUnsafePointer(moduleHash[:]),
+		unsafe.Pointer(gasPtr),
+		arbutil.SliceToUnsafePointer(errBuf),
+		uint32(len(errBuf)),
 	)
-	if err != nil {
-		_, _, err := userFailure.toResult(err.intoSlice(), debug)
+	if errLen != 0 {
+		err := errors.New(string(errBuf[:errLen]))
 		return nil, err
 	}
 	return &activationInfo{moduleHash, initGas, asmEstimate, footprint}, nil
 }
+
+//go:wasmimport programs new_program
+func newProgram(
+	hashPtr unsafe.Pointer,
+	callDataPtr unsafe.Pointer,
+	callDataSize uint32,
+	configHandler stylusConfigHandler,
+	evmHandler evmDataHandler,
+	gas uint64,
+) uint32
+
+//go:wasmimport programs pop
+func popProgram()
+
+//go:wasmimport programs set_response
+func setResponse(id uint32, gas uint64, result unsafe.Pointer, result_len uint32, raw_data unsafe.Pointer, raw_data_len uint32)
+
+//go:wasmimport programs get_request
+func getRequest(id uint32, reqLen unsafe.Pointer) uint32
+
+//go:wasmimport programs get_request_data
+func getRequestData(id uint32, dataPtr unsafe.Pointer)
+
+//go:wasmimport programs start_program
+func startProgram(module uint32) uint32
+
+//go:wasmimport programs send_response
+func sendResponse(req_id uint32) uint32
 
 func callProgram(
 	address common.Address,
@@ -93,46 +128,40 @@ func callProgram(
 	params *goParams,
 	memoryModel *MemoryModel,
 ) ([]byte, error) {
-	debug := arbmath.UintToBool(params.debugMode)
-	evmApi := newApi(interpreter, tracingInfo, scope, memoryModel)
-	defer evmApi.drop()
+	// debug := arbmath.UintToBool(params.debugMode)
+	reqHandler := newApiClosures(interpreter, tracingInfo, scope, memoryModel)
 
-	status, output := callProgramRustImpl(
-		&moduleHash,
-		calldata,
-		params.encode(),
-		evmApi.id,
-		evmData.encode(),
-		&scope.Contract.Gas,
+	configHandler := params.createHandler()
+	dataHandler := evmData.createHandler()
+
+	module := newProgram(
+		unsafe.Pointer(&moduleHash[0]),
+		arbutil.SliceToUnsafePointer(calldata),
+		uint32(len(calldata)),
+		configHandler,
+		dataHandler,
+		scope.Contract.Gas,
 	)
-	data, _, err := status.toResult(output.intoSlice(), debug)
-	return data, err
-}
-
-func (vec *rustVec) intoSlice() []byte {
-	len := readRustVecLenImpl(vec)
-	slice := make([]byte, len)
-	rustVecIntoSliceImpl(vec, arbutil.SliceToPointer(slice))
-	return slice
-}
-
-func (p *goParams) encode() *rustConfig {
-	return rustConfigImpl(p.version, p.maxDepth, p.inkPrice.ToUint32(), p.debugMode)
-}
-
-func (d *evmData) encode() *rustEvmData {
-	return rustEvmDataImpl(
-		&d.blockBasefee,
-		u64(d.chainId),
-		&d.blockCoinbase,
-		u64(d.blockGasLimit),
-		u64(d.blockNumber),
-		u64(d.blockTimestamp),
-		&d.contractAddress,
-		&d.msgSender,
-		&d.msgValue,
-		&d.txGasPrice,
-		&d.txOrigin,
-		u32(d.reentrant),
-	)
+	reqId := startProgram(module)
+	for {
+		var reqLen uint32
+		reqTypeId := getRequest(reqId, unsafe.Pointer(&reqLen))
+		reqData := make([]byte, reqLen)
+		getRequestData(reqId, arbutil.SliceToUnsafePointer(reqData))
+		if reqTypeId < EvmApiMethodReqOffset {
+			popProgram()
+			status := userStatus(reqTypeId)
+			gasLeft := binary.BigEndian.Uint64(reqData[:8])
+			scope.Contract.Gas = gasLeft
+			data, msg, err := status.toResult(reqData[8:], params.debugMode != 0)
+			if status == userFailure && params.debugMode != 0 {
+				log.Warn("program failure", "err", err, "msg", msg, "program", address)
+			}
+			return data, err
+		}
+		reqType := RequestType(reqTypeId - EvmApiMethodReqOffset)
+		result, rawData, cost := reqHandler(reqType, reqData)
+		setResponse(reqId, cost, arbutil.SliceToUnsafePointer(result), uint32(len(result)), arbutil.SliceToUnsafePointer(rawData), uint32(len(rawData)))
+		reqId = sendResponse(reqId)
+	}
 }
