@@ -1,8 +1,8 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-//go:build !js
-// +build !js
+//go:build !wasm
+// +build !wasm
 
 package wavmio
 
