@@ -3,6 +3,7 @@
 
 use crate::{evm::user::UserOutcomeKind, Bytes20, Bytes32};
 use eyre::Result;
+use std::sync::Arc;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[repr(u8)]
@@ -30,7 +31,7 @@ impl From<u8> for EvmApiStatus {
 }
 
 #[derive(Clone, Copy, Debug)]
-#[repr(usize)]
+#[repr(u32)]
 pub enum EvmApiMethod {
     GetBytes32,
     SetBytes32,
@@ -39,16 +40,41 @@ pub enum EvmApiMethod {
     StaticCall,
     Create1,
     Create2,
-    GetReturnData,
     EmitLog,
     AccountBalance,
     AccountCode,
-    AccountCodeSize,
     AccountCodeHash,
     AddPages,
+    CaptureHostIO,
 }
 
-pub trait EvmApi: Send + 'static {
+// This offset is added to EvmApiMethod when sending a request
+// in WASM - program done is also indicated by a "request", with the
+// id below that offset, indicating program status
+pub const EVM_API_METHOD_REQ_OFFSET: u32 = 0x10000000;
+
+// note: clone should not clone actual data, just the reader
+pub trait DataReader: Clone + Send + 'static {
+    fn slice(&self) -> &[u8];
+}
+
+// simple implementation for DataReader, in case data comes from a Vec
+#[derive(Clone, Debug)]
+pub struct VecReader(Arc<Vec<u8>>);
+
+impl VecReader {
+    pub fn new(data: Vec<u8>) -> Self {
+        Self(Arc::new(data))
+    }
+}
+
+impl DataReader for VecReader {
+    fn slice(&self) -> &[u8] {
+        self.0.as_slice()
+    }
+}
+
+pub trait EvmApi<D: DataReader>: Send + 'static {
     /// Reads the 32-byte value in the EVM state trie at offset `key`.
     /// Returns the value and the access cost in gas.
     /// Analogous to `vm.SLOAD`.
@@ -115,7 +141,7 @@ pub trait EvmApi: Send + 'static {
 
     /// Returns the EVM return data.
     /// Analogous to `vm.RETURNDATASIZE`.
-    fn get_return_data(&mut self, offset: u32, size: u32) -> Vec<u8>;
+    fn get_return_data(&self) -> D;
 
     /// Emits an EVM log with the given number of topics and data, the first bytes of which should be the topic data.
     /// Returns an error message on failure.
@@ -129,17 +155,7 @@ pub trait EvmApi: Send + 'static {
 
     /// Returns the code and the access cost in gas.
     /// Analogous to `vm.EXTCODECOPY`.
-    fn account_code(
-        &mut self,
-        address: Bytes20,
-        offset: u32,
-        size: u32,
-        gas_left: u64,
-    ) -> (Vec<u8>, u64);
-
-    /// Returns the code size and the access cost in gas.
-    /// Analogous to `vm.EXTCODESIZE`.
-    fn account_code_size(&mut self, address: Bytes20, gas_left: u64) -> (u32, u64);
+    fn account_code(&mut self, address: Bytes20, gas_left: u64) -> (D, u64);
 
     /// Gets the hash of the given address's code.
     /// Returns the hash and the access cost in gas.
@@ -152,5 +168,12 @@ pub trait EvmApi: Send + 'static {
     fn add_pages(&mut self, pages: u16) -> u64;
 
     /// Captures tracing information for hostio invocations during native execution.
-    fn capture_hostio(&self, name: &str, args: &[u8], outs: &[u8], start_ink: u64, end_ink: u64);
+    fn capture_hostio(
+        &mut self,
+        name: &str,
+        args: &[u8],
+        outs: &[u8],
+        start_ink: u64,
+        end_ink: u64,
+    );
 }
