@@ -2,6 +2,7 @@
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
 use eyre::ErrReport;
+use num_enum::{IntoPrimitive, TryFromPrimitive};
 use std::fmt::Display;
 
 #[derive(Debug)]
@@ -13,7 +14,7 @@ pub enum UserOutcome {
     OutOfStack,
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, TryFromPrimitive, IntoPrimitive)]
 #[repr(u8)]
 pub enum UserOutcomeKind {
     Success,
@@ -56,12 +57,6 @@ impl From<&UserOutcome> for UserOutcomeKind {
 impl From<&UserOutcome> for u8 {
     fn from(value: &UserOutcome) -> Self {
         UserOutcomeKind::from(value).into()
-    }
-}
-
-impl From<UserOutcomeKind> for u8 {
-    fn from(value: UserOutcomeKind) -> Self {
-        value as u8
     }
 }
 
