@@ -6,18 +6,17 @@ use crate::machine::{Escape, WasmEnv};
 use arbutil::{color, Color};
 use eyre::Result;
 use structopt::StructOpt;
-use wasmer::Value;
 
 use std::path::PathBuf;
 
 mod arbcompress;
-mod gostack;
+mod callerenv;
 mod machine;
-mod runtime;
+mod program;
 mod socket;
-mod syscall;
+mod stylus_backend;
 mod test;
-mod user;
+mod wasip1_stub;
 mod wavmio;
 
 #[derive(StructOpt)]
@@ -46,8 +45,6 @@ pub struct Opts {
     #[structopt(long)]
     forks: bool,
     #[structopt(long)]
-    go_arg: bool,
-    #[structopt(long)]
     debug: bool,
     #[structopt(long)]
     require_success: bool,
@@ -62,26 +59,9 @@ fn main() -> Result<()> {
     };
 
     let (instance, env, mut store) = machine::create(&opts, env);
-    let mut run_args = vec![Value::I32(0), Value::I32(0)];
 
-    if opts.go_arg {
-        let memory = instance.exports.get_memory("mem").unwrap();
-        let memory = memory.view(&store);
-
-        // To pass in the program name argument, we need to put it in memory.
-        // The Go linker guarantees a section of memory starting at byte 4096 is available for this purpose.
-        // https://github.com/golang/go/blob/252324e879e32f948d885f787decf8af06f82be9/misc/wasm/wasm_exec.js#L520
-        let free_memory_base: i32 = 4096;
-        let name = free_memory_base;
-        let argv = name + 8;
-
-        memory.write(name as u64, b"js\0")?; // write "js\0" to the name ptr
-        memory.write(argv as u64, &name.to_le_bytes())?; // write the name ptr to the argv ptr
-        run_args = vec![Value::I32(1), Value::I32(argv)]; // pass argv with our single name arg
-    }
-
-    let main = instance.exports.get_function("run").unwrap();
-    let outcome = main.call(&mut store, &run_args);
+    let main = instance.exports.get_function("_start").unwrap();
+    let outcome = main.call(&mut store, &vec![]);
     let escape = match outcome {
         Ok(outcome) => {
             println!("Go returned values {:?}", outcome);
