@@ -1,13 +1,14 @@
 // Copyright 2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
+//use core::slice::SlicePattern;
 use std::{
     io,
     io::{BufReader, BufWriter, Read, Write},
     net::TcpStream,
 };
 
-use crate::wavmio::Bytes32;
+use arbutil::Bytes32;
 
 pub const SUCCESS: u8 = 0x0;
 pub const FAILURE: u8 = 0x1;
@@ -31,8 +32,8 @@ pub fn read_u64<T: Read>(reader: &mut BufReader<T>) -> Result<u64, io::Error> {
 }
 
 pub fn read_bytes32<T: Read>(reader: &mut BufReader<T>) -> Result<Bytes32, io::Error> {
-    let mut buf = Bytes32::default();
-    reader.read_exact(&mut buf).map(|_| buf)
+    let mut buf = [0u8; 32];
+    reader.read_exact(&mut buf).map(|_| buf.into())
 }
 
 pub fn read_bytes<T: Read>(reader: &mut BufReader<T>) -> Result<Vec<u8>, io::Error> {
@@ -57,7 +58,7 @@ pub fn write_u64(writer: &mut BufWriter<TcpStream>, data: u64) -> Result<(), io:
 }
 
 pub fn write_bytes32(writer: &mut BufWriter<TcpStream>, data: &Bytes32) -> Result<(), io::Error> {
-    writer.write_all(data)
+    writer.write_all(data.as_slice())
 }
 
 pub fn write_bytes(writer: &mut BufWriter<TcpStream>, data: &[u8]) -> Result<(), io::Error> {
