@@ -78,7 +78,6 @@ pub enum Hostio {
     WavmHaltAndSetFinished,
     WavmLinkModule,
     WavmUnlinkModule,
-    WavmSetErrorPolicy,
     ProgramInkLeft,
     ProgramInkStatus,
     ProgramSetInk,
@@ -86,6 +85,8 @@ pub enum Hostio {
     ProgramSetStack,
     ProgramMemorySize,
     ProgramCallMain,
+    ProgramRequest,
+    ProgramContinue,
     ConsoleLogTxt,
     ConsoleLogI32,
     ConsoleLogI64,
@@ -122,7 +123,6 @@ impl FromStr for Hostio {
             ("env", "wavm_halt_and_set_finished") => WavmHaltAndSetFinished,
             ("hostio", "wavm_link_module") => WavmLinkModule,
             ("hostio", "wavm_unlink_module") => WavmUnlinkModule,
-            ("hostio", "wavm_set_error_policy") => WavmSetErrorPolicy,
             ("hostio", "program_ink_left") => ProgramInkLeft,
             ("hostio", "program_ink_status") => ProgramInkStatus,
             ("hostio", "program_set_ink") => ProgramSetInk,
@@ -130,6 +130,8 @@ impl FromStr for Hostio {
             ("hostio", "program_set_stack") => ProgramSetStack,
             ("hostio", "program_memory_size") => ProgramMemorySize,
             ("hostio", "program_call_main") => ProgramCallMain,
+            ("hostio", "program_request") => ProgramRequest,
+            ("hostio", "program_continue") => ProgramContinue,
             ("hostio", "user_ink_left") => UserInkLeft,
             ("hostio", "user_ink_status") => UserInkStatus,
             ("hostio", "user_set_ink") => UserSetInk,
@@ -180,7 +182,6 @@ impl Hostio {
             WavmHaltAndSetFinished      => func!(),
             WavmLinkModule              => func!([I32], [I32]),      // λ(module_hash) → module
             WavmUnlinkModule            => func!(),                  // λ()
-            WavmSetErrorPolicy          => func!([I32]),             // λ(enabled)
             ProgramInkLeft              => func!([I32], [I64]),      // λ(module) → ink_left
             ProgramInkStatus            => func!([I32], [I32]),      // λ(module) → ink_status
             ProgramSetInk               => func!([I32, I64]),        // λ(module, ink_left)
@@ -188,6 +189,8 @@ impl Hostio {
             ProgramSetStack             => func!([I32, I32]),        // λ(module, stack_left)
             ProgramMemorySize           => func!([I32], [I32]),      // λ(module) → memory_size
             ProgramCallMain             => func!([I32, I32], [I32]), // λ(module, args_len) → status
+            ProgramRequest              => func!([I32], [I32]),      // λ(status) → response
+            ProgramContinue             => func!([I32], [I32]), // λ(response) → status
             ConsoleLogTxt               => func!([I32, I32]),        // λ(text, len)
             ConsoleLogI32               => func!([I32]),             // λ(value)
             ConsoleLogI64               => func!([I64]),             // λ(value)
@@ -291,15 +294,12 @@ impl Hostio {
                 // λ(module_hash) → module
                 opcode!(LocalGet, 0);
                 opcode!(LinkModule);
+                opcode!(NewCoThread);
             }
             WavmUnlinkModule => {
                 // λ()
                 opcode!(UnlinkModule);
-            }
-            WavmSetErrorPolicy => {
-                // λ(status)
-                opcode!(LocalGet, 0);
-                opcode!(SetErrorPolicy);
+                opcode!(PopCoThread);
             }
             ProgramInkLeft => {
                 // λ(module) → ink_left
@@ -329,16 +329,42 @@ impl Hostio {
                 cross_internal!(UserMemorySize);
             }
             ProgramCallMain => {
-                // λ(module, args_len) → status
-                opcode!(PushErrorGuard);
-                opcode!(ArbitraryJumpIf, prior + body.len() + 3);
-                opcode!(I32Const, UserOutcomeKind::Failure as u32);
+                // caller sees: λ(module, args_len) → status
+                opcode!(LocalGet, 1); // args_len
+                opcode!(LocalGet, 0); // module
+                opcode!(MoveFromStackToInternal);
+                opcode!(MoveFromStackToInternal);
+                opcode!(SwitchThread, 8);
+                opcode!(MoveFromInternalToStack);
+                opcode!(MoveFromInternalToStack);
+                opcode!(CrossModuleInternalCall, InternalFunc::CallMain); // consumes module
+                opcode!(MoveFromStackToInternal);
+                opcode!(SwitchThread, 0);
+                opcode!(MoveFromInternalToStack);
                 opcode!(Return);
 
-                // jumps here in the happy case
-                opcode!(LocalGet, 1); // args_len
-                cross_internal!(CallMain);
-                opcode!(PopErrorGuard);
+                // jumps here if errored while in thread 1
+                opcode!(I32Const, UserOutcomeKind::Failure as u32)
+            }
+            ProgramContinue => {
+                // caller sees: λ(return_data) → status (returns to caller of ProgramRequest)
+                // code returns return_data to caller of ProgramRequest
+                opcode!(LocalGet, 0); // return_data
+                opcode!(MoveFromStackToInternal);
+                opcode!(SwitchThread, 3);
+                opcode!(MoveFromInternalToStack);
+                opcode!(Return);
+
+                // jumps here if errored while in cothread
+                opcode!(I32Const, UserOutcomeKind::Failure as u32)
+            }
+            ProgramRequest => {
+                // caller sees: λ(status)->response
+                // code returns status of either ProgramContinue or ProgramCallMain
+                opcode!(LocalGet, 0); // return_data
+                opcode!(MoveFromStackToInternal);
+                opcode!(SwitchThread, 0);
+                opcode!(MoveFromInternalToStack);
             }
             UserInkLeft => {
                 // λ() → ink_left
