@@ -1,12 +1,11 @@
 // Copyright 2022-2023, Offchain Labs, Inc.
 // For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
-
-use crate::evm_api::ApiCaller;
+use core::sync::atomic::{compiler_fence, Ordering};
 use arbutil::{
-    evm::{js::JsEvmApi, EvmData},
+    evm::{req::{EvmApiRequestor, RequestHandler}, EvmData, api::{{VecReader, EvmApiMethod, EVM_API_METHOD_REQ_OFFSET}}},
     wavm, Bytes20, Bytes32, Color,
 };
-use eyre::{eyre, Result};
+use eyre::{bail, eyre, Result};
 use prover::programs::prelude::*;
 use std::fmt::Display;
 use user_host_trait::UserHost;
@@ -37,6 +36,27 @@ impl From<MemoryBoundsError> for eyre::ErrReport {
 #[allow(clippy::vec_box)]
 static mut PROGRAMS: Vec<Box<Program>> = vec![];
 
+static mut LAST_REQUEST_ID: u32 = 0x10000;
+
+#[derive(Clone)]
+pub (crate) struct UserHostRequester {
+    data: Option<Vec<u8>>,
+    answer: Option<(Vec<u8>, VecReader, u64)>,
+    req_type: u32,
+    id: u32,
+}
+
+impl UserHostRequester {
+    pub fn default() -> Self {
+        Self {
+            req_type: 0,
+            data: None,
+            answer: None,
+            id: 0,
+        }
+    }
+}
+
 /// An active user program.
 pub(crate) struct Program {
     /// Arguments passed via the VM.
@@ -44,7 +64,7 @@ pub(crate) struct Program {
     /// Output generated by the program.
     pub outs: Vec<u8>,
     /// Mechanism for calling back into Geth.
-    pub evm_api: JsEvmApi<ApiCaller>,
+    pub evm_api: EvmApiRequestor<VecReader, UserHostRequester>,
     /// EVM Context info.
     pub evm_data: EvmData,
     /// WAVM module index.
@@ -53,11 +73,69 @@ pub(crate) struct Program {
     pub config: StylusConfig,
 }
 
+#[link(wasm_import_module = "hostio")]
+extern "C" {
+    fn program_request(status: u32) -> u32;
+}
+
+impl UserHostRequester {
+    #[no_mangle]
+    pub unsafe fn set_response(&mut self, req_id: u32, result: Vec<u8>, raw_data:Vec<u8>, gas: u64) {
+        self.answer = Some((result, VecReader::new(raw_data), gas));
+        if req_id != self.id {
+            panic!("bad req id returning from send_request")
+        }
+        compiler_fence(Ordering::SeqCst);
+    }
+
+    pub unsafe fn set_request(&mut self, req_type: u32, data: &[u8]) -> u32 {
+        LAST_REQUEST_ID += 1;
+        self.id = LAST_REQUEST_ID;
+        self.req_type = req_type;
+        self.data = Some(data.to_vec());
+        self.answer = None;
+        self.id
+    }
+
+    pub unsafe fn get_request_meta(&self, id: u32) -> (u32, usize) {
+        if self.id != id {
+            panic!("get_request got wrong id");
+        }
+        (self.req_type, self.data.as_ref().expect("no request on get_request_meta").len())
+    }
+
+    pub unsafe fn take_request(&mut self, id: u32) -> (u32, Vec<u8>) {
+        if self.id != id {
+            panic!("get_request got wrong id");
+        }
+        (self.req_type, self.data.take().expect("no request on take_request"))
+    }
+
+    #[no_mangle]
+    unsafe fn send_request(&mut self, req_type: u32, data: Vec<u8>) -> (Vec<u8>, VecReader, u64) {
+        let req_id = self.set_request(req_type, &data);
+        compiler_fence(Ordering::SeqCst);
+        let got_id = program_request(req_id);
+        compiler_fence(Ordering::SeqCst);
+        if got_id != req_id {
+            panic!("bad req id returning from send_request")
+        }
+        self.answer.take().unwrap()
+    }
+}
+
+impl RequestHandler<VecReader> for UserHostRequester {
+    fn handle_request(&mut self, req_type: EvmApiMethod, req_data: &[u8]) -> (Vec<u8>, VecReader, u64) {
+        unsafe {
+            self.send_request(req_type as u32 + EVM_API_METHOD_REQ_OFFSET, req_data.to_vec())
+        }
+    }
+}
+
 impl Program {
     /// Adds a new program, making it current.
     pub fn push_new(
         args: Vec<u8>,
-        evm_api: JsEvmApi<ApiCaller>,
         evm_data: EvmData,
         module: u32,
         config: StylusConfig,
@@ -65,7 +143,7 @@ impl Program {
         let program = Self {
             args,
             outs: vec![],
-            evm_api,
+            evm_api: EvmApiRequestor::new(UserHostRequester::default()),
             evm_data,
             module,
             config,
@@ -73,9 +151,9 @@ impl Program {
         unsafe { PROGRAMS.push(Box::new(program)) }
     }
 
-    /// Removes the current program, returning its output.
-    pub fn pop() -> Vec<u8> {
-        unsafe { PROGRAMS.pop().expect("no program").outs }
+    /// Removes the current program
+    pub fn pop() {
+        unsafe { PROGRAMS.pop().expect("no program"); }
     }
 
     /// Provides a reference to the current program.
@@ -86,6 +164,11 @@ impl Program {
     /// Reads the program's memory size in pages
     fn memory_size(&self) -> u32 {
         unsafe { program_memory_size(self.module) }
+    }
+
+    /// Reads the program's memory size in pages
+    pub fn args_len(&self) -> usize {
+        return self.args.len()
     }
 
     /// Ensures an access is within bounds
@@ -99,10 +182,10 @@ impl Program {
 }
 
 #[allow(clippy::unit_arg)]
-impl UserHost for Program {
+impl UserHost<VecReader> for Program {
     type Err = eyre::ErrReport;
     type MemoryErr = MemoryBoundsError;
-    type A = JsEvmApi<ApiCaller>;
+    type A = EvmApiRequestor<VecReader, UserHostRequester>;
 
     fn args(&self) -> &[u8] {
         &self.args
@@ -163,9 +246,13 @@ impl UserHost for Program {
         println!("{} {text}", "Stylus says:".yellow());
     }
 
-    fn trace(&self, name: &str, args: &[u8], outs: &[u8], _end_ink: u64) {
+    fn trace(&mut self, name: &str, args: &[u8], outs: &[u8], _start_ink: u64, _end_ink: u64) {
         let args = hex::encode(args);
         let outs = hex::encode(outs);
         println!("Error: unexpected hostio tracing info for {name} while proving: {args}, {outs}");
+    }
+
+    fn start_ink(&self) -> Result<u64,Self::Err> {
+        bail!("recording start ink while proving")
     }
 }
